FastAPI_backend/requirements.txt
facenet_pytorch==2.5.3
fastapi==0.101.1
joblib==1.3.2
Pillow==10.0.0
torch==2.0.1
torchvision==0.15.2
uvicorn==0.23.2
scikit_learn==1.3.0
numpy==1.24.4
python-multipart==0.0.6
redis==5.0.0
